Move some of the fields in ScDocument into a thread-local struct
This is part of the work to enable having several ScInterpreters
working on the same object in parallel. Perhaps a more thorugh
re-factoring would be better, maybe these fields should not be part of
ScDocument at all, but part of ScInterpreter?

Without this change the parallelism unit test occasionally fails.

Change-Id: I70500b4d2b513e134d22b4b98c84a3ec00dad291